functional; fixed bug that was causing some words to not load
